chore: move dataset registration page
