+++ src/components/dataset-registration-form/dataset-registration-page.scss
@@ -1,5 +1,10 @@
 @import '../../../node_modules/designsystemet/fdk-designsystem-bootstrap4/scss/colors.scss';
 
+.dataset-regitration-form-title {
+  font-size: 2.8rem;
+  font-weight: 600;
+}
+
 .fdk-reg_backgroundDefault {
   background-color: $color-white;
   padding: 1.6rem 3.2rem;
@@ -16,11 +21,6 @@
   }
 }
 
-.chevronIcon {
-  height: 20px;
-  width: 20px;
-}
-
 .toggleExpandButton {
   border: none;
   color: $color-link;
@@ -28,6 +28,11 @@
   padding: 8px 8px;
   cursor: pointer;
   display: inline-block;
+
+  & > svg {
+    height: 20px;
+    width: 20px;
+  }
 }
 
 .form-group {
