--- src/pages/dataset-registration-page/dataset-registration-connector.js
@@ -1,72 +1,0 @@
-import { connect } from 'react-redux';
-import _ from 'lodash';
-
-import {
-  deleteDatasetThunk,
-  selectorForDatasetsInCatalog
-} from '../../entrypoints/main/redux/modules/datasets';
-import {
-  setInputLanguages,
-  toggleInputLanguage
-} from '../../components/language-picker/redux/actions';
-import {
-  REFERENCEDATA_PATH_FREQUENCY,
-  REFERENCEDATA_PATH_LOS,
-  REFERENCEDATA_PATH_OPENLICENCES,
-  REFERENCEDATA_PATH_PROVENANCE,
-  REFERENCEDATA_PATH_REFERENCETYPES,
-  REFERENCEDATA_PATH_MEDIATYPES,
-  REFERENCEDATA_PATH_THEMES
-} from '../../entrypoints/main/redux/modules/referenceData';
-import { selectorForDatasetFormStatus } from '../../entrypoints/main/redux/modules/dataset-form-status';
-import { datasetRegistrationEnsureDataThunk } from './dataset-registration-ensure-data-thunk';
-
-const mapStateToProps = (state, { catalogId, datasetId }) => {
-  const {
-    form,
-    referenceData,
-    inputLanguage: { languages }
-  } = state;
-  const datasetFormStatus = selectorForDatasetFormStatus(datasetId)(state);
-  const referenceDatasetsItems = Object.values(
-    selectorForDatasetsInCatalog(catalogId)(state)
-  );
-
-  return {
-    form,
-    datasetFormStatus,
-    provenanceItems: _.get(referenceData, [
-      'items',
-      REFERENCEDATA_PATH_PROVENANCE
-    ]),
-    frequencyItems: _.get(referenceData, [
-      'items',
-      REFERENCEDATA_PATH_FREQUENCY
-    ]),
-    themesItems: _.get(referenceData, ['items', REFERENCEDATA_PATH_THEMES]),
-    referenceTypesItems: _.get(referenceData, [
-      'items',
-      REFERENCEDATA_PATH_REFERENCETYPES
-    ]),
-    referenceDatasetsItems,
-    openLicenseItems: _.get(referenceData, [
-      'items',
-      REFERENCEDATA_PATH_OPENLICENCES
-    ]),
-    mediaTypes: _.get(referenceData, ['items', REFERENCEDATA_PATH_MEDIATYPES]),
-    losItems: _.get(referenceData, ['items', REFERENCEDATA_PATH_LOS]),
-    languages
-  };
-};
-
-const mapDispatchToProps = {
-  dispatchEnsureData: datasetRegistrationEnsureDataThunk,
-  dispatchDeleteDataset: deleteDatasetThunk,
-  setInputLanguages,
-  toggleInputLanguage
-};
-
-export const datasetRegistrationConnector = connect(
-  mapStateToProps,
-  mapDispatchToProps
-);
